+++ RedPandaIDE/src/settingsdialog/editorcodecompletionwidget.cpp
@@ -49,7 +49,6 @@ void EditorCodeCompletionWidget::doLoad()
     ui->chkIgnoreCases->setChecked(pSettings->codeCompletion().ignoreCase());
     ui->chkAppendFunc->setChecked(pSettings->codeCompletion().appendFunc());
     ui->chkShowCodeIns->setChecked(pSettings->codeCompletion().showCodeIns());
-    ui->chkClearWhenEditorHidden->setChecked(pSettings->codeCompletion().clearWhenEditorHidden());
     ui->chkHideSymbolsStartWithTwoUnderline->setChecked(pSettings->codeCompletion().hideSymbolsStartsWithTwoUnderLine());
     ui->chkHideSymbolsStartWithUnderline->setChecked(pSettings->codeCompletion().hideSymbolsStartsWithUnderLine());
 
@@ -73,8 +72,6 @@ void EditorCodeCompletionWidget::doSave()
     pSettings->codeCompletion().setAppendFunc(ui->chkAppendFunc->isChecked());
     pSettings->codeCompletion().setShowCodeIns(ui->chkShowCodeIns->isChecked());
     pSettings->codeCompletion().setMinCharRequired(ui->spinMinCharRequired->value());
-
-    pSettings->codeCompletion().setClearWhenEditorHidden(ui->chkClearWhenEditorHidden->isChecked());
 
     pSettings->codeCompletion().setHideSymbolsStartsWithTwoUnderLine(ui->chkHideSymbolsStartWithTwoUnderline->isChecked());
     pSettings->codeCompletion().setHideSymbolsStartsWithUnderLine(ui->chkHideSymbolsStartWithUnderline->isChecked());