+++ RedPandaIDE/src/settingsdialog/environmentperformancewidget.cpp
@@ -33,14 +33,12 @@ EnvironmentPerformanceWidget::~EnvironmentPerformanceWidget()
 
 void EnvironmentPerformanceWidget::doLoad()
 {
-    ui->chkClearWhenEditorHidden->setChecked(pSettings->codeCompletion().clearWhenEditorHidden());
     ui->chkEditorsShareParser->setChecked(pSettings->codeCompletion().shareParser());
     on_chkEditorsShareParser_stateChanged(false);
 }
 
 void EnvironmentPerformanceWidget::doSave()
 {
-    pSettings->codeCompletion().setClearWhenEditorHidden(ui->chkClearWhenEditorHidden->isChecked());
     pSettings->codeCompletion().setShareParser(ui->chkEditorsShareParser->isChecked());
 
     pSettings->codeCompletion().save();
@@ -50,6 +48,5 @@ void EnvironmentPerformanceWidget::doSave()
 void EnvironmentPerformanceWidget::on_chkEditorsShareParser_stateChanged(int arg1)
 {
     Q_UNUSED(arg1)
-    ui->chkClearWhenEditorHidden->setVisible(ui->chkEditorsShareParser->isChecked());
 }