+++ RedPandaIDE/test/resources/struct.cpp
@@ -1,0 +1,9 @@
+struct TTT {
+	int xxx;
+	short yyy;
+	double zzz;
+private:
+	int ppp;
+}
+
+