+++ RedPandaIDE/test/resources/structured-bindings.cpp
@@ -1,0 +1,13 @@
+struct T {
+	int xxx;
+	short yyy;
+	double zzz;
+};
+
+T f(int x);
+
+auto [a,b,c]=T();
+auto [d,e,g]=f(5);
+
+
+